Fix dimensionality issues
This was done by multiplying setting output channel to double of
input channel, so as to achieve "same" padding as in keras, i.e
input shape and output shape remains the same. Also changing padding
in the conv net to make the output height and width dimens the
same as well.